requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
babel==2.17.0
beautifulsoup4==4.13.4
bing-image-downloader==1.1.2
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
curated-tokenizers==2.0.0
curated-transformers==2.0.1
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
dm-tree==0.1.9
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==8.0.4
einops==0.8.1
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
et_xmlfile==2.0.0
etils==1.12.2
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
ImageHash==4.3.2
imageio==2.37.0
immutabledict==4.2.1
importlib_resources==6.5.2
imutils==0.5.4
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
isoduration==20.11.0
jax==0.4.34
jaxlib==0.4.34
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
kagglehub==0.3.12
keras==3.9.2
keras-core==0.1.7
keras-cv==0.9.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml-dtypes==0.3.2
mlxtend==0.23.4
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
namex==0.0.9
narwhals==1.40.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
overrides==7.7.0
packaging==24.2
panda==0.3.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
platformdirs==4.3.8
plotly==6.1.1
prefetch_generator==1.0.3
preshed==3.0.9
primp==0.15.0
prometheus_client==0.22.0
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.21.12
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
roboflow==1.2.2
rpds-py==0.25.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
spacy==3.8.7
spacy-alignments==0.9.1
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.8
split-folders==0.5.1
srsly==2.5.1
stack-data==0.6.3
streamlit==1.45.1
streamlit-webrtc==0.63.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.8
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metadata==1.17.1
tensorflow-metal==1.2.0
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.144
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
zipp==3.21.0
